# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "SynergyVAHouse API Documentation",
    "version": "1.0.0",
    "description": "# SynergyVAHouse API\n\nA comprehensive task management and workflow automation platform with webhook integrations, AI assistance, and RBAC.\n\n## Authentication\n\nThe API supports multiple authentication methods:\n\n### 1. JWT Bearer Tokens\n```\nAuthorization: Bearer <your_jwt_token>\n```\n\n### 2. API Keys\n```\nX-API-Key: <your_api_key>\n```\n\n### 3. Webhook Signatures\nFor webhook endpoints (Conduit and SuiteOp):\n```\nX-Conduit-Signature: sha256=<hmac_signature>\nX-SuiteOp-Signature: sha256=<hmac_signature>\n```\n\n### 4. Slack Signatures\nFor Slack events:\n```\nX-Slack-Signature: v0=<hmac_signature>\nX-Slack-Request-Timestamp: <unix_timestamp>\n```\n\n## Rate Limiting\n\nAPI endpoints are rate-limited to ensure fair usage. Please implement exponential backoff for retries.\n\n## Idempotency\n\nWebhook endpoints support idempotency through event IDs. Duplicate events with the same ID will return a 200 status with `status: 'duplicate'`.",
    "contact": {
      "name": "SynergyVAHouse Team"
    },
    "license": {
      "name": "Private"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "API Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from /api/auth/login"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key for service-to-service authentication"
      },
      "webhookSignature": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Conduit-Signature",
        "description": "HMAC-SHA256 signature for webhook verification"
      },
      "slackSignature": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Slack-Signature",
        "description": "Slack request signature verification"
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message"
          },
          "details": {
            "type": "object",
            "description": "Additional error details"
          }
        },
        "required": [
          "error"
        ]
      },
...
    },
...
  },
...
  "paths": {
...
    "/webhooks/suiteop": {
      "post": {
        "tags": [
          "Webhooks"
        ],
        "summary": "SuiteOp webhook endpoint",
        "description": "Receives webhook events from SuiteOp for task creation and updates.\n\nSupported event types:\n- `task.created` - Creates a new reactive task from SuiteOp\n- `task.updated` - Updates an existing task status\n\nSecurity:\n- HMAC-SHA256 signature verification via X-SuiteOp-Signature header\n- Idempotency via event ID tracking (prevents duplicate processing)\n- Event ID extraction: X-Event-ID header > body.id > body.event_id > SHA256 hash\n\nFeatures:\n- Automatic Definition of Done (DoD) schema population from playbooks\n- SLA timer initialization when playbook is available\n- Audit logging for all events\n- Status mapping: completed → DONE, in_progress → IN_PROGRESS, cancelled → BLOCKED\n",
        "security": [
          {
            "webhookSignature": []
          }
        ],
        "parameters": [
          {
            "in": "header",
            "name": "X-SuiteOp-Signature",
            "required": true,
            "schema": {
              "type": "string",
              "example": "sha256=a1b2c3d4e5f6..."
            },
            "description": "HMAC-SHA256 signature of the request body"
          },
          {
            "in": "header",
            "name": "X-Event-ID",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Optional unique event identifier for idempotency"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "task.created",
                      "task.updated"
                    ],
                    "description": "Event type"
                  },
                  "id": {
                    "type": "string",
                    "description": "Event ID (used for idempotency if X-Event-ID not provided)"
                  },
                  "event_id": {
                    "type": "string",
                    "description": "Alternative event ID field"
                  },
                  "task": {
                    "type": "object",
                    "description": "Task data",
                    "properties": {
                      "id": {
                        "type": "string",
                        "description": "SuiteOp task ID"
                      },
                      "title": {
                        "type": "string",
                        "description": "Task title"
                      },
                      "description": {
                        "type": "string",
                        "description": "Task description"
                      },
                      "status": {
                        "type": "string",
                        "enum": [
                          "pending",
                          "in_progress",
                          "completed",
                          "cancelled"
                        ],
                        "description": "Task status"
                      },
                      "priority": {
                        "type": "integer",
                        "minimum": 1,
                        "maximum": 5,
                        "description": "Task priority (1=highest, 5=lowest)"
                      },
                      "url": {
                        "type": "string",
                        "format": "uri",
                        "description": "Link to task in SuiteOp"
                      },
                      "category": {
                        "type": "string",
                        "description": "Task category/type"
                      }
                    }
                  }
                }
              },
              "examples": {
                "task_created": {
                  "summary": "Task Created Event",
                  "value": {
                    "type": "task.created",
                    "id": "evt_123456789",
                    "task": {
                      "id": "task_987654321",
                      "title": "Process refund request",
                      "description": "Customer requested refund for order #12345",
                      "status": "pending",
                      "priority": 2,
                      "category": "refund",
                      "url": "https://app.suiteop.com/tasks/987654321"
                    }
                  }
                },
                "task_updated": {
                  "summary": "Task Updated Event",
                  "value": {
                    "type": "task.updated",
                    "id": "evt_123456790",
                    "task": {
                      "id": "task_987654321",
                      "status": "completed"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Webhook processed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": [
                        "processed",
                        "duplicate"
                      ],
                      "description": "Processing status"
                    },
                    "taskId": {
                      "type": "string",
                      "format": "uuid",
                      "description": "ID of created/updated task (if applicable)"
                    },
                    "eventId": {
                      "type": "string",
                      "description": "Event ID (only present for duplicate events)"
                    },
                    "message": {
                      "type": "string",
                      "description": "Human-readable message (only for duplicates)"
                    },
                    "processedAt": {
                      "type": "string",
                      "format": "date-time",
                      "description": "When event was originally processed (only for duplicates)"
                    }
                  }
                },
                "examples": {
                  "processed": {
                    "summary": "Successfully Processed",
                    "value": {
                      "status": "processed",
                      "taskId": "550e8400-e29b-41d4-a716-446655440000"
                    }
                  },
                  "duplicate": {
                    "summary": "Duplicate Event",
                    "value": {
                      "status": "duplicate",
                      "message": "Event already processed",
                      "eventId": "evt_123456789",
                      "processedAt": "2025-10-22T12:00:00Z",
                      "taskId": "550e8400-e29b-41d4-a716-446655440000"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid JSON payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Invalid JSON payload"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or missing HMAC signature",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Invalid signature"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error during processing",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Processing failed"
                }
              }
            }
          }
        }
      }
    },
...
  }
}